+++ addon/mixins/run.js
@@ -47,7 +47,6 @@ export default Mixin.create({
     this._pendingTimers = undefined;
     this._pendingDebounces = undefined;
     this._pollerLabels = undefined;
-    this._registeredDisposables = undefined;
   },
 
   /**
@@ -368,107 +367,12 @@ export default Mixin.create({
     cancelPoll(label);
   },
 
-  /**
-   Adds a new disposable to the object. A disposable is a function that
-   disposes of bindings that are outside of Ember's lifecyle. This essentially
-   means you can register a function that you want to run to automatically tear
-   down any bindings when the Ember object is destroyed.
-
-   Example:
-
-   ```js
-   // app/components/foo-bar.js
-   import Ember from 'ember';
-   import DOMish from 'some-external-library';
-
-   const { run } = Ember;
-
-   export default Component.extend({
-     init() {
-       this.DOMish = new DOMish();
-
-       this.bindEvents();
-     },
-
-     bindEvents() {
-       let onFoo = run.bind(this.respondToDomEvent);
-       this.DOMish.on('foo', onFoo);
-
-       this.domFooToken = this.registerDisposable(() => this.DOMish.off('foo', onFoo));
-     },
-
-     respondToDOMEvent() {
-       // do something
-     }
-   });
-   ```
-
-   @method registerDisposable
-   @param { Function } disposable
-   @returns A label representing the position of the disposable
-   @public
-   */
-  registerDisposable(disposable) {
-    if (typeof disposable !== 'function') {
-      throw new Error('You must pass a function as a disposable');
-    }
-
-    let disposables = this._getOrAllocateArray('_registeredDisposables');
-
-    return disposables.push(disposable) - 1;
-  },
-
-  /**
-   Runs a disposable identified by the supplied token.
-
-  ```js
-   // app/components/foo-bar.js
-   import DOMish from 'some-external-library';
-   import Ember from 'ember';
-
-   const { run } = Ember;
-
-   export default Component.extend({
-     init() {
-       this.DOMish = new DOMish();
-
-       this.bindEvents();
-     },
-
-     bindEvents() {
-       let onFoo = run.bind(this.respondToDomEvent);
-       this.DOMish.on('foo', onFoo);
-
-       this.domFooToken = this.registerDisposable(() => this.DOMish.off('foo', onFoo));
-     },
-
-     respondToDOMEvent() {
-       // do something
-     },
-
-     actions: {
-       cancelDOM() {
-         this.runDisposable(this.domFooToken);
-       }
-     }
-   });
-   ```
-
-   @method runDisposable
-   @param {any} label
-   @public
-   */
-  runDisposable(token) {
-    runDisposable(this._registeredDisposables, token);
-  },
-
   willDestroy() {
     this._super(...arguments);
 
     cancelTimers(this._pendingTimers);
     cancelDebounces(this._pendingDebounces);
     clearPollers(this._pollerLabels);
-    runDisposables(this._registeredDisposables);
   },
 
   _getOrAllocateArray(propertyName) {
@@ -532,29 +436,4 @@ function cancelDebounces(pendingDebounces) {
 function cancelDebounce(pendingDebounces, name) {
   let { cancelId } = pendingDebounces[name];
   run.cancel(cancelId);
-}
-
-function runDisposables(disposables) {
-  if (!disposables) {
-    return;
-  }
-
-  for (let i = 0, l = disposables.length; i < l; i++) {
-    let disposable = disposables.pop();
-
-    disposable();
-  }
-}
-
-function runDisposable(disposables, token) {
-  if (!disposables || token < 0) {
-    return;
-  }
-
-  let disposable = disposables[token];
-
-  if (disposable) {
-    disposables.splice(token, 1);
-    disposable();
-  }
 }
