+++ testem.js
@@ -8,12 +8,12 @@ module.exports = {
   launch_in_dev: [
     'Chrome'
   ],
-  'browser_args': {
-      'Chrome': [
-        '--disable-gpu',
-        '--headless',
-        '--remote-debugging-port=9222',
-        '--window-size=1440,900'
-      ],
-  }
+  // 'browser_args': {
+  //     'Chrome': [
+  //       '--disable-gpu',
+  //       '--headless',
+  //       '--remote-debugging-port=9222',
+  //       '--window-size=1440,900'
+  //     ],
+  // }
 };
