+++ tests/unit/mixins/run-test.js
@@ -494,61 +494,3 @@ test('pollTask: can be manually cleared', function(assert) {
       pollTaskFor('one');
     });
 });
-
-test('registerDisposable: ensures a function is passed as a disposable', function(assert) {
-  assert.expect(1);
-
-  let subject = this.subject();
-
-  assert.throws(() => {
-    subject.registerDisposable({});
-  }, /You must pass a function as a disposable/);
-});
-
-test('registerDisposable: returns a label when a disposable is registered', function(assert) {
-  assert.expect(2);
-
-  let subject = this.subject();
-  let disposable = () => {};
-
-  let label = subject.registerDisposable(disposable);
-
-  assert.equal(label, 0, 'label is returned');
-
-  let otherLabel = subject.registerDisposable(disposable);
-
-  assert.notEqual(label, otherLabel, 'label returned is unique');
-});
-
-test('runDisposable: runs the disposable when called with valid label', function(assert) {
-  assert.expect(1);
-
-  let callCount = 0;
-  let subject = this.subject();
-  let disposable = () => {
-    callCount++;
-  };
-
-  let label = subject.registerDisposable(disposable);
-
-  subject.runDisposable(label);
-
-  assert.equal(callCount, 1, 'disposable is called');
-});
-
-test('runDisposables: runs all disposables when destroying', function(assert) {
-  assert.expect(2);
-
-  let subject = this.subject();
-  let disposable = () => {};
-  let disposableTheSecond = () => {};
-
-  subject.registerDisposable(disposable);
-  subject.registerDisposable(disposableTheSecond);
-
-  assert.equal(subject._registeredDisposables.length, 2, 'two disposables are registered');
-
-  run(subject, 'destroy');
-
-  assert.equal(subject._registeredDisposables.length, 0, 'no disposables are registered');
-});
